% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_rel_affected.R
\name{risk_rel_affected}
\alias{risk_rel_affected}
\title{Estimate Prevalence of Disease Among d-Degree Relatives}
\usage{
risk_rel_affected(q, K, r2, d)
}
\arguments{
\item{q}{(numeric; range = 0 to 1) Top quantile of risk used to define "elevated" risk in the proband. For example, if interested in the top 1\% of risk, q = 0.01.}

\item{K}{(numeric; range = 0 to 1) Population prevalence of disease}

\item{r2}{(numeric; range = 0 to 1) Proportion of variability explained by the PRS (r^2^)}

\item{d}{(integer) Degree of relationship to relatives}
}
\description{
This function enables the user to estimate the prevalence of disease among \code{d}-degree relatives of probands with elevated polygenic risk (within the top \code{q} of the PRS distribution), based on a disease with population prevalence \code{K}, and proportion of variability (r^2^) explained by the PRS \code{r2}.
}
\examples{
risk_rel_affected(q = 0.25, d = 1, r2 = 0.02, K = 0.05)
}
