% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_rel_high_prs.R
\name{risk_rel_high_prs}
\alias{risk_rel_high_prs}
\title{Estimate Prevalence of Elevated PRS Among Relatives of d-degree}
\usage{
risk_rel_high_prs(q, d)
}
\arguments{
\item{q}{(numeric; range = 0 to 1) Top quantile of risk used to define "elevated" risk in the proband. For example, if interested in the top 1\% of risk, q = 0.01.}

\item{d}{(integer) Degree of relationship to relatives}
}
\value{
(numeric) Prevalence of elevated polygenic risk (top \code{q} of the PRS distribution) among relatives of \code{d}-degree
}
\description{
This function enables the user to estimate the prevalence of elevated polygenic risk among \code{d}-degree relatives of probands with elevated polygenic risk (within the top \code{q} of the PRS distribution).
}
\examples{
risk_rel_high_prs(q = 0.1, d = 1)
}
