% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_sib_affected.R
\name{risk_sib_affected}
\alias{risk_sib_affected}
\title{Estimate Prevalence of Disease Among First-Degree Relatives}
\usage{
risk_sib_affected(q, K, r)
}
\arguments{
\item{q}{(numeric; range = 0 to 1) Top quantile of risk used to define "elevated" risk in the proband. For example, if interested in the top 1\% of risk, q = 0.01.}

\item{K}{(numeric; range = 0 to 1) Population prevalence of disease}

\item{r}{(numeric; range = 0 to 1) Proportion of variability explained by the PRS (r^2^)}
}
\value{
(numeric) Prevalence of disease among relatives
}
\description{
This function enables the user to estimate the prevalence of disease among first-degree relatives of probands with elevated polygenic risk (within the top \code{q} of the PRS distribution), based on a disease with population prevalence \code{K}, and proportion of variability (r^2^) explained by the PRS \code{r}.
}
\examples{
risk_sib_affected(q = 0.01, K = 0.05, r = 0.1)
}
