% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_sib_high_prs.R
\name{risk_sib_high_prs}
\alias{risk_sib_high_prs}
\title{Estimate Prevalence of Elevated Polygenic Risk Among First-Degree Relatives}
\usage{
risk_sib_high_prs(q)
}
\arguments{
\item{q}{(numeric; range = 0 to 1) Top quantile of risk used to define "elevated" risk in the proband. For example, if interested in the top 1\% of risk, q = 0.01.}
}
\value{
(numeric) Prevalence of elevated polygenic risk among first-degree relatives
}
\description{
This function enables the user to estimate the prevalence of elevated polygenic risk among first-degree relatives of probands with elevated polygenic risk (within the top \code{q} of the PRS distribution).
}
\examples{
risk_sib_high_prs(q = .25)
}
